capi: support all Flatcar Container Linux channels
This change adds support for building all channels of Flatcar
Container Linux. It also fixes an issue that made the build fail
when make was called directly instead of using the wrapper script
hack/image-build-flatcar.sh.

Define an env variable `CAPI_PRIVIDER`, which is `aws`, `ami` or `qemu`.
Default is `qemu`.

Run vagrant commands only when the provider is qemu.
If it is `aws` or `ami`, run `make build-ami-default`.
Otherwise fail.